requirements.txt
# Core Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database & ORM
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
alembic==1.13.1

# Data Validation & Settings
pydantic==2.11.7
pydantic-settings==2.2.1
pydantic[email]

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Email
fastapi-mail==1.4.1

# Other Utilities
python-multipart==0.0.9
redis==5.0.4